Split regularizations into general and mesh-based
Strip volumes out of weights in mesh-based regularizations. Add a new
`Smallness` regularization.
