+++ src/inversion_ideas/regularization/__init__.py
@@ -1,0 +1,11 @@
+"""
+Regularization classes.
+"""
+from ._general import TikhonovZero
+from ._mesh_based import Smallness, Smoothness
+
+__all__ = [
+    "Smallness",
+    "Smoothness",
+    "TikhonovZero",
+]
